requirements.txt
# Adaptive Chessbot Dependencies

# Chess libraries
python-chess>=1.999

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Frontend
streamlit>=1.28.0

# HTTP client
requests>=2.31.0

# Data manipulation for plotting
pandas>=2.1.0

# Environment variables
python-dotenv>=1.0.0

# For type checking
pydantic>=2.5.0

# Testing
pytest>=7.4.0
matplotlib
